+++ lib/quds_db.dart
@@ -2,6 +2,8 @@
 library sqlite_quds_db;
 
 // import 'dart:ffi';
+import 'dart:convert';
+import 'dart:ffi';
 import 'dart:io';
 import 'dart:typed_data';
 import 'package:flutter/cupertino.dart';
@@ -17,7 +19,6 @@ part 'quds_db/data_page_query.dart';
 part 'quds_db/data_page_query_result.dart';
 part 'quds_db/db_functions.dart';
 part 'quds_db/db_helper.dart';
-// part 'quds_db/_web_sql_helper.dart';
 part 'quds_db/db_model.dart';
 part 'quds_db/db_table_provider.dart';
 part 'quds_db/entry_change_type.dart';
@@ -30,6 +31,7 @@ part 'quds_db/query_parts/order_field.dart';
 part 'quds_db/query_parts/operator_query.dart';
 
 //Db fields
+part 'quds_db/db_fields/enum_field.dart';
 part 'quds_db/db_fields/blob_field.dart';
 part 'quds_db/db_fields/bool_field.dart';
 part 'quds_db/db_fields/color_field.dart';
@@ -38,5 +40,6 @@ part 'quds_db/db_fields/double_field.dart';
 part 'quds_db/db_fields/int_field.dart';
 part 'quds_db/db_fields/num_field.dart';
 part 'quds_db/db_fields/string_field.dart';
+part 'quds_db/db_fields/json_field.dart';
 part 'quds_db/db_fields/id_field.dart';
 part 'quds_db/db_fields/datetime_string_field.dart';