+++ CHANGELOG.md
@@ -23,3 +23,12 @@
 
 ## [0.0.3+4]
 * Added some documentations
+
+## [0.0.4]
+* Support multiple tables for same provider
+
+## [0.0.4+1]
+* Add Support to save enums as db fields
+
+## [0.0.4+2]
+* Add Support to save json data in the db